requirements.txt
python-jenkins==1.8.1
requests==2.31.0
PyYAML==6.0.1
